Only make legal bottle moves
The previous commit won't actually run with asserts on since the planner was still planning illegal bottle moves.

This fixes the planner to consider an order of bottles to potentially move (ideally it probably should create full plans for each bottle and compare how many spaces each could actually use, etc.).

I'm not actually sure all this logic is 100% correct, but this does drive the win rate another 3% up to 50%.

This has a lot of testing, but still probably needs more.

I've seen numerous games which seem to be lost for lack of reveals.  I'm suspicious that the reveal logic isn't good enough or maybe I just need to compute and print # of magics rolled in lost games.