Add basic flutter UI
I also added a coordinate system (in polar-ish coordinates) for Spaces, to make them easier to index and draw.

removed widget_test.dart, since #1 it's now broken (with removal of the counter logic) but also I haven't figured out how to have both dart and flutter tests coexist in the same repo.
There are still UI bugs about not knowing how to draw multiple tokens on a space correctly, etc.  But there is so much easy to build on top of this.